[1/2] base: battery styles
BatteryViewManager knows all the different battery styles and
will switch based on settings. No more need to mess with the
global layouts.

All required to add a new style is to add it to BatteryViewManager
and to the resource array of the styles

Start with basic bar and circle styles and optional battery percent support.
More can be added over time

Change-Id: If44dda883a1a53796a0eda54cd208f0b0802df01

Conflicts:
	core/java/android/provider/Settings.java
	core/res/res/values/custom_arrays.xml
	core/res/res/values/custom_strings.xml
	core/res/res/values/custom_symbols.xml